# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gabinete Digital Cícero João
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      tags: [Public]
      operationId: listProjects
      summary: List projects and laws
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [Admin]
      operationId: createProject
      summary: Create a project / law (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{id}:
    get:
      tags: [Public]
      operationId: getProject
      summary: Get a single project / law
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Admin]
      operationId: updateProject
      summary: Update a project / law (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: deleteProject
      summary: Delete a project / law (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /news:
    get:
      tags: [Public]
      operationId: listNews
      summary: List news articles
      responses:
        "200":
          description: News
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      tags: [Admin]
      operationId: createNewsArticle
      summary: Create a news article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "201":
          description: News article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /news/{id}:
    get:
      tags: [Public]
      operationId: getNewsArticle
      summary: Get a single news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Admin]
      operationId: updateNewsArticle
      summary: Update a news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleUpdate"
      responses:
        "200":
          description: News article updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: deleteNewsArticle
      summary: Delete a news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /appointment-slots:
    get:
      tags: [Citizen]
      operationId: listAvailableSlots
      summary: List released appointment slots with remaining availability
      responses:
        "200":
          description: Released slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentSlot"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/appointment-slots:
    get:
      tags: [Admin]
      operationId: listSlots
      summary: List all appointment slots (staff)
      responses:
        "200":
          description: Slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentSlot"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Admin]
      operationId: createSlot
      summary: Create an appointment slot card (manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentSlotInput"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentSlot"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/appointment-slots/{id}:
    patch:
      tags: [Admin]
      operationId: updateSlot
      summary: Release/hold or adjust an appointment slot card (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentSlotUpdate"
      responses:
        "200":
          description: Slot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentSlot"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Project:
      type: object
      required: [id, title, summary, type, status]
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        theme:
          type: ["string", "null"]
        type:
          type: string
          description: "projeto_lei | requerimento | indicacao | emenda | lei"
        status:
          type: string
          description: "protocolado | em_tramitacao | aprovado | sancionado | arquivado"
        year:
          type: ["integer", "null"]
        documentUrl:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
          format: date-time
    ProjectInput:
      type: object
      required: [title, summary, type, status]
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        theme:
          type: string
        type:
          type: string
        status:
          type: string
        year:
          type: integer
        documentUrl:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        theme:
          type: string
        type:
          type: string
        status:
          type: string
        year:
          type: integer
        documentUrl:
          type: string
    NewsArticle:
      type: object
      required: [id, title, summary, body, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        body:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
          format: date-time
    NewsArticleInput:
      type: object
      required: [title, summary, body]
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        imageUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
    NewsArticleUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        imageUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
...
    AppointmentSlot:
      type: object
      required: [id, date, period, capacity, bookedCount, released, available, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        period:
          type: string
          description: "manha | tarde"
        capacity:
          type: integer
        bookedCount:
          type: integer
        released:
          type: boolean
        available:
          type: integer
          description: "remaining capacity (capacity - bookedCount)"
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentSlotInput:
      type: object
      required: [date, period]
      properties:
        date:
          type: string
          format: date
        period:
          type: string
        capacity:
          type: integer
        released:
          type: boolean
        note:
          type: ["string", "null"]
    AppointmentSlotUpdate:
      type: object
      properties:
        capacity:
          type: integer
        released:
          type: boolean
        note:
          type: ["string", "null"]
...